requirements.txt
pytest-bdd==3.2.1
requests==2.22.0
django==3.0.4
djangorestframework==3.11.0